requirements.txt
python-dotenv==1.0.1
schedule==1.2.2
slack_bolt==1.20.0
slack_sdk==3.31.0
